+++ packages/ice/src/utils/prepareURLs.ts
@@ -37,33 +37,36 @@ export default function prepareUrls(
   let lanUrlForTerminal: string;
   if (isUnspecifiedHost) {
     prettyHost = 'localhost';
-    try {
-      // This can only return an IPv4 address
-      lanUrlForConfig = address.ip();
-      if (lanUrlForConfig) {
-        // Check if the address is a private ip
-        // https://en.wikipedia.org/wiki/Private_network#Private_IPv4_address_spaces
-        if (
-          /^10[.]|^30[.]|^172[.](1[6-9]|2[0-9]|3[0-1])[.]|^192[.]168[.]/.test(
-            lanUrlForConfig,
-          ) ||
-          process.env.USE_PUBLIC_IP
-        ) {
-          // Address is private, format it for later use
-          lanUrlForTerminal = prettyPrintUrl(lanUrlForConfig);
-        } else {
-          // Address is not private, so we will discard it
-          lanUrlForConfig = undefined;
-        }
-      }
-    } catch (_e) {
-      // ignored
-    }
   } else {
     prettyHost = host;
   }
+
+  try {
+    // This can only return an IPv4 address
+    lanUrlForConfig = address.ip();
+    if (lanUrlForConfig) {
+      // Check if the address is a private ip
+      // https://en.wikipedia.org/wiki/Private_network#Private_IPv4_address_spaces
+      if (
+        /^10[.]|^30[.]|^172[.](1[6-9]|2[0-9]|3[0-1])[.]|^192[.]168[.]/.test(
+          lanUrlForConfig,
+        ) ||
+        process.env.USE_PUBLIC_IP
+      ) {
+        // Address is private, format it for later use
+        lanUrlForTerminal = prettyPrintUrl(lanUrlForConfig);
+      } else {
+        // Address is not private, so we will discard it
+        lanUrlForConfig = undefined;
+      }
+    }
+  } catch (_e) {
+    // ignored
+  }
+
   const localUrlForTerminal = prettyPrintUrl(prettyHost);
   const localUrlForBrowser = formatUrl(prettyHost);
+
   return {
     lanUrlForConfig,
     lanUrlForTerminal,