+++ packages/ice/src/createService.ts
@@ -21,7 +21,7 @@ import { setEnv, updateRuntimeEnv, getCoreEnvKeys } from './utils/runtimeEnv.js'
 import getRuntimeModules from './utils/getRuntimeModules.js';
 import { generateRoutesInfo } from './routes.js';
 import * as config from './config.js';
-import { RUNTIME_TMP_DIR, WEB } from './constant.js';
+import { RUNTIME_TMP_DIR, WEB, RUNTIME_EXPORTS } from './constant.js';
 import createSpinner from './utils/createSpinner.js';
 import getRoutePaths from './utils/getRoutePaths.js';
 import ServerCompileTask from './utils/ServerCompileTask.js';
@@ -100,6 +100,11 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
     plugins.push(pluginWeb());
   }
 
+  // Register framework level API.
+  RUNTIME_EXPORTS.forEach(exports => {
+    generatorAPI.addExport(exports);
+  });
+
   const ctx = new Context<Config, ExtendsPluginAPI>({
     rootDir,
     command,
@@ -147,44 +152,6 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
     ctx.registerConfig(configType, configData);
   });
   let taskConfigs = await ctx.setup();
-
-  // Register framework level API.
-  generatorAPI.addExport({
-    specifier: ['Link', 'Outlet', 'useParams', 'useSearchParams', 'useLocation', 'useNavigate'],
-    source: '@ice/runtime/router',
-  });
-
-  generatorAPI.addExport({
-    specifier: [
-      'defineAppConfig',
-      'useAppData',
-      'useData',
-      'useConfig',
-      'history',
-      'KeepAliveOutlet',
-      'useMounted',
-      'ClientOnly',
-      'withSuspense',
-      'useSuspenseData',
-      'defineDataLoader',
-      'defineServerDataLoader',
-      'defineStaticDataLoader',
-    ],
-    source: '@ice/runtime',
-  });
-
-  generatorAPI.addExport({
-    specifier: [
-      'Meta',
-      'Title',
-      'Links',
-      'Scripts',
-      'Data',
-      'Main',
-    ],
-    source: '@ice/runtime',
-    target: 'web',
-  });
 
   // get userConfig after setup because of userConfig maybe modified by plugins
   const { userConfig } = ctx;