+++ packages/ice/src/constant.ts
@@ -1,7 +1,9 @@
 import * as path from 'path';
 
+export const DEFAULT_PORT = 3000;
+export const DEFAULT_HOST = '0.0.0.0';
+
 export const RUNTIME_TMP_DIR = '.ice';
-export const ROUTER_MANIFEST = path.join(RUNTIME_TMP_DIR, 'route-manifest.json');
 export const SERVER_ENTRY = path.join(RUNTIME_TMP_DIR, 'entry.server.ts');
 export const DATA_LOADER_ENTRY = path.join(RUNTIME_TMP_DIR, 'data-loader.ts');
 export const SERVER_OUTPUT_DIR = 'server';