+++ packages/ice/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## v3.1.1
 
-- [fix] The imported dataloaderFetcher name is error
 - [fix] pass external and define info for preBundle
+- [fix] the lanUrlForConfig should set when set host
+- [fix] the imported dataloaderFetcher name is error
+- [fix] the value of `isServer` is false when compile data-loader
 
 ## v3.1.0