+++ packages/ice/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v3.1.1
 
+- [feat] redirect runtime imports for data loader
 - [fix] the lanUrlForConfig should set when set host
 - [fix] the imported dataloaderFetcher name is error
 - [fix] the value of `isServer` is false when compile data-loader